changes addressed
